--- src/main/java/com/packsendme/microservice/sa/component/AirwayConsumer_Component.java
@@ -1,19 +1,0 @@
-package com.packsendme.microservice.sa.component;
-
-public class AirwayConsumer_Component {
-
-	private String airwayMsg; 
-	
-	/*
-	@KafkaListener(topics = "${kafka.topic.airwaybre}")
-	public void receiveTopic(String msg) {
-		this.airwayMsg = msg;
-		System.out.println(" ---------------------------- ");
-		System.out.println(" topic_airway_sa "+ msg);
-		System.out.println(" ---------------------------- ");
-	} */
-
-	public String contextMsg() {
-		return airwayMsg;
-	}
-}
