--- src/main/java/com/packsendme/microservice/sa/component/BREConsumer_Interface.java
@@ -1,8 +1,0 @@
-package com.packsendme.microservice.sa.component;
-
-public interface BREConsumer_Interface {
-	
-	public void receiveTopic(String msg);
-	public String contextMsg();
-
-}
